+++ test/PackageSvgLinkTest.php
@@ -37,11 +37,9 @@ final class PackageSvgLinkTest extends TestCase
         foreach ($files as $f) {
             $filename = sprintf('%s/namespace/%s', $this->fixtureDir, $f);
             $classes = PhpReader::parseFile($directory, $filename, $options);
-            foreach ($classes as $c) {
-                $entries = array_merge($entries, [new Entry('product', $c->getInfo(), $options)]);
-            }
+            $entries[] = array_map(fn($c) => new Entry(dirname($f), $c->getInfo(), $options), $classes);
         }
-        $rel = new Relation($entries, $options);
+        $rel = new Relation(array_merge(...$entries), $options);
 
         $expected = <<<EOS
 @startuml class-diagram
@@ -75,11 +73,9 @@ EOS;
         foreach ($files as $f) {
             $filename = sprintf('%s/%s', $directory, $f);
             $classes = PhpReader::parseFile($directory, $filename, $options);
-            foreach ($classes as $c) {
-                $entries = array_merge($entries, [new Entry(dirname($f), $c->getInfo(), $options)]);
-            }
+            $entries[] = array_map(fn($c) => new Entry(dirname($f), $c->getInfo(), $options), $classes);
         }
-        $rel = new Relation($entries, $options);
+        $rel = new Relation(array_merge(...$entries), $options);
         $expected = <<<EOS
 @startuml class-diagram
   skinparam svgLinkTarget _blank
@@ -110,11 +106,9 @@ EOS;
         foreach ($files as $f) {
             $filename = sprintf('%s/%s', $directory, $f);
             $classes = PhpReader::parseFile($directory, $filename, $options);
-            foreach ($classes as $c) {
-                $entries = array_merge($entries, [new Entry(dirname($f), $c->getInfo(), $options)]);
-            }
+            $entries[] = array_map(fn($c) => new Entry(dirname($f), $c->getInfo(), $options), $classes);
         }
-        $rel = new Relation($entries, $options);
+        $rel = new Relation(array_merge(...$entries), $options);
 
         $expected = <<<EOS
 @startuml class-diagram